requirements.txt
azure-common==1.1.27
azure-core==1.21.1
azure-servicebus==7.5.0
certifi==2021.10.8
charset-normalizer==2.0.10
idna==3.3
isodate==0.6.1
msrest==0.6.21
oauthlib==3.1.1
requests==2.27.1
requests-oauthlib==1.3.0
six==1.16.0
typing-extensions==4.0.1
uamqp==1.5.1
urllib3==1.26.8
